+++ deployment/ConfigASDK.ps1
@@ -40,6 +40,11 @@
     * Supports usage in offline/disconnected environments
 
 .VERSION
+    1906.1  Minor bug fixes and cleanup
+            App Service File Server deployment improvements
+            App Service File Server now uses Server Core 2016 image
+            Updated Ubuntu Server 16.04 Image - 16.04.20190628
+            Servicing Stack Update now dynamically gathered at runtime
     1906    Updated to support ASDK 1.1906.0.30
             Moved backend RP resources to Tenant Space
             Minor bug fixes and cleanup